Set sale price changes

+++ src/main/java/com/mapiz/mystore/stock/application/command/SetSalePriceToStockProductCommand.java
@@ -1,5 +1,6 @@
 package com.mapiz.mystore.stock.application.command;
 
+import java.math.BigDecimal;
 import lombok.Builder;
 import lombok.Getter;
 
@@ -7,5 +8,5 @@ import lombok.Getter;
 @Getter
 public class SetSalePriceToStockProductCommand {
   private int productId;
-  private int salePrice;
+  private BigDecimal salePrice;
 }
