+++ src/main/java/com/mapiz/mystore/stock/application/dto/StockItemSummary.java
@@ -12,7 +12,7 @@ public class StockItemSummary {
   private Integer productId;
   private String productName;
   private BigDecimal quantity;
-  private double salePrice;
+  private BigDecimal salePrice;
   private BigDecimal weightedCost;
 
   public void sumQuantity(BigDecimal quantity) {
